move EditingMenu component out of Geral subfolder
