+++ website/src/Components/FileSystem/Geral/FileSystem.js
@@ -26,7 +26,7 @@ import Notification from 'Components/Notifications/Geral/Notification';
 import FullStorageMenu from "Components/Notifications/Geral/FullStorageMenu";
 import OcrMenu from 'Components/OcrMenu/Geral/OcrMenu';
 import LayoutMenu from 'Components/LayoutMenu/Geral/LayoutMenu';
-import EditingMenu from 'Components/EditingMenu/Geral/EditingMenu';
+import EditingMenu from 'Components/EditingMenu/EditingMenu';
 import FolderMenu from 'Components/Form/Geral/FolderMenu';
 import OcrPopup from 'Components/Form/Geral/OcrPopup';
 import DeletePopup from 'Components/Form/Geral/DeletePopup';
